Moved over files from previous repo.
The only new thing different from the other repo is the CabinetUIShim, which is an AutoHotKey script that waits on the running process to close before closing itself. It also allows you to kill the currently running program with a button on the cabinet (which is the actual reason why it exists).