Added shim launcher intergration.
This comes as two functions. One launches the game directly (as a control) and the other launches the game using the shim. Both use the same arguments, and both should have similar behavior. The shim is assumed to be in the Games folder (which itself is assumed to be in the Data folder).
This code needs testing on Windows. It probably won't work without any modifications.